requirements.txt
os==
psycopg2==2.9.9
PyPDF2==3.0.1
logging==0.4.9.6
python-dotenv==1.0.0